tokenize comments and skip them in the parsing
This adds an explicit lookahead in the statement parser, compared to the
current unbounded lookahead. rustc complains if we try to match
something that requires more lookahead, but it won't complain if we have
too much lookahead.

We need the explicit lookahead since the old &ctx.tokens[ctx.curr..]
would otherwise contain comments.